Fixing file structure issues in charger.

+++ Charger_Control.X/Functions.h
@@ -2,13 +2,13 @@
  * File:   Functions.h
  * Author: Rick
  *
- * Created on October 17, 2015, 12:48 AM
+ * Created on May 11, 2015, 2:42 AM
  */
 
 #ifndef FUNCTIONS_H
 #define	FUNCTIONS_H
 
-extern void Delay(int wait);
+extern void Delay(long int wait);
 
 #endif	/* FUNCTIONS_H */
